requirements.txt
accelerate==0.20.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==41.0.1
diffusers==0.29.0
filelock==3.9.0
firebase-admin==6.1.0
fsspec==2024.6.0
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
huggingface-hub==0.23.4
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
msgpack==1.0.5
networkx==3.0rc1
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
proto-plus==1.22.2
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.28.1
rsa==4.9
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sympy==1.11.1
tokenizers==0.19.1
torch==2.1.0.dev20230616
torchaudio==2.1.0.dev20230616
torchvision==0.16.0.dev20230616
tqdm==4.65.0
transformers==4.41.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.15.0
